% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HettiePikeBCI-package.R
\docType{package}
\name{HettiePikeBCI-package}
\alias{HettiePikeBCI-package}
\alias{HettiePikeBCI}
\title{Barro Colorado Island data package}
\description{
Package to hold the BCI data (or whatever) – maybe also mention something
about these functions now, and put that in the DESCRIPTION too. And then
put it in the README.md file. And don't forget to reference the source of
the data correctly.
}
\details{
\code{HettiePikeBCI} package developed as part of the Reproducible
Programming in R course.
}
\author{
Hettie Pike
}
