% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HettiePikeBCI-package.R
\docType{package}
\name{HettiePikeBIC-package}
\alias{HettiePikeBIC-package}
\alias{HettiePikeBCI}
\title{HettiePikeBCI}
\description{
\code{HettiePikeBCI} package developed as part of the Reproducible
Programming in R course.
}
\author{
is Hettie
}
